# openapi.yaml
openapi: 3.0.3
info:
  title: ExecFlex API
  description: |
    ExecFlex API - Combined API Server for executive matching and voice interactions.
    
    This API handles both web-based operations (role posting, matching, introductions) 
    and voice/telephony features powered by Ai-dan, an AI voice agent.
    
    ## Features
    - Executive candidate matching
    - Role posting and management
    - Email introductions
    - Voice calls via Twilio (Ai-dan)
    
    ## Authentication
    Most endpoints require JWT authentication via Supabase Auth. Include the JWT token in the Authorization header:
    `Authorization: Bearer <supabase_jwt_token>`
    
    Twilio webhook endpoints (`/voice/*`) are public and called by Twilio (signature verification recommended for production).
  version: 1.0.0
  contact:
    name: ExecFlex API Support
    url: https://execflex.com
servers:
  - url: https://execflex-backend-1.onrender.com
    description: Production server (Render)
  - url: http://localhost:5001
    description: Local development server
...
paths:
...
  /voice/intro:
    post:
      tags:
        - Voice
      summary: Twilio webhook - Inbound call entry point
      description: |
        Entry point for inbound voice calls from users.
        Called by Twilio when a user calls the ExecFlex phone number.
        
        **Note:** This endpoint is for inbound calls only. For outbound qualification calls,
        see `/voice/qualify`.
        
        **Note:** This is a webhook endpoint, called by Twilio, not directly by API clients.
      operationId: voiceIntro
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                CallSid:
                  type: string
                  description: Twilio Call SID
      responses:
        '200':
          description: TwiML response for inbound call handling
          content:
            application/xml:
              schema:
                type: string
                description: TwiML XML with opening message and Gather
        '503':
          description: Voice features not available
    get:
      tags:
        - Voice
      summary: Twilio webhook - Inbound call entry point (GET)
      description: Alternative GET handler for inbound calls
      operationId: voiceIntroGet
      responses:
        '200':
          description: TwiML response
          content:
            application/xml:
              schema:
                type: string
        '503':
          description: Voice features not available
  /voice/qualify:
    post:
      tags:
        - Voice
      summary: Twilio webhook - Outbound qualification call handler
      description: |
        Outbound qualification call endpoint. Called by Twilio when an outbound qualification call is answered.
        
        This endpoint handles all logic for qualification calls:
        - OpenAI reasoning (with job seeker or talent seeker system prompts)
        - Storing interaction + turns in database
        - Converting text to speech with ElevenLabs
        - Incremental DB updates (people_profiles, organizations, role_assignments)
        
        **Flow:**
        1. Initial call: Twilio calls this endpoint when user answers → Returns opening message + Gather
        2. Subsequent turns: Twilio calls this endpoint after Gather → Processes user speech, calls OpenAI, returns next message
        
        **Note:** This is a webhook endpoint, called by Twilio, not directly by API clients.
        **Security:** Twilio signature verification required (X-Twilio-Signature header).
      operationId: voiceQualify
      parameters:
        - name: job_id
          in: query
          description: Job ID from outbound_call_jobs table (required)
          required: true
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                CallSid:
                  type: string
                  description: Twilio Call SID
                  required: true
                SpeechResult:
                  type: string
                  description: Recognized speech text from user (present only for subsequent turns)
                Confidence:
                  type: string
                  description: Speech recognition confidence score (0.0-1.0)
      responses:
        '200':
          description: TwiML response for next turn or call termination
          content:
            application/xml:
              schema:
                type: string
                description: TwiML XML with next message, Gather (if continuing), or Hangup (if complete)
        '400':
          description: Missing job_id
        '403':
          description: Invalid Twilio signature
        '503':
          description: Voice features not available
    get:
      tags:
        - Voice
      summary: Twilio webhook - Outbound qualification call handler (GET)
      description: Alternative GET handler
      operationId: voiceQualifyGet
      parameters:
        - name: job_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: TwiML response
          content:
            application/xml:
              schema:
                type: string
        '503':
          description: Voice features not available
...
  /voice/status:
    post:
      tags:
        - Voice
      summary: Unified Twilio webhook - Call status callback
      description: |
        Unified status callback webhook for all Twilio voice calls.
        Receives call status updates (initiated, ringing, answered, completed, failed, etc.)
        and updates job and interaction records accordingly.
        
        **Note:** This is a webhook endpoint, called by Twilio, not directly by API clients.
        **Security:** Twilio signature verification required (X-Twilio-Signature header).
      operationId: voiceStatus
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                CallSid:
                  type: string
                  description: Twilio Call SID
                  required: true
                CallStatus:
                  type: string
                  description: Call status
                  enum: [queued, ringing, in-progress, completed, failed, busy, no-answer, canceled]
                CallDuration:
                  type: string
                  description: Call duration in seconds (only for completed calls)
                From:
                  type: string
                  description: Caller's phone number
                To:
                  type: string
                  description: Called phone number
      responses:
        '200':
          description: Status update processed
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        '400':
          description: Missing CallSid
        '403':
          description: Invalid Twilio signature
  /voice/capture:
    post:
      tags:
        - Voice
      summary: Legacy Twilio webhook - Speech capture (deprecated)
      description: |
        **DEPRECATED:** Legacy endpoint for inbound call conversation flow.
        Kept for backward compatibility only.
        
        Use `/voice/inbound` instead for inbound call handling.
        
        **Note:** This is a webhook endpoint, typically called by Twilio, not directly by API clients.
      operationId: voiceCapture
      deprecated: true
      parameters:
        - name: step
          in: query
          description: Current conversation step
          schema:
            type: string
            enum: [user_type, name, role, industry, location, availability, confirm_intro, email]
          example: "name"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                CallSid:
                  type: string
                  description: Twilio Call SID
                SpeechResult:
                  type: string
                  description: Recognized speech text
                Confidence:
                  type: string
                  description: Speech recognition confidence score
      responses:
        '200':
          description: TwiML response for next step or call termination
          content:
            application/xml:
              schema:
                type: string
                description: TwiML XML
        '503':
          description: Voice features not available
    get:
      tags:
        - Voice
      summary: Legacy Twilio webhook - Speech capture (GET, deprecated)
      description: Alternative GET handler (deprecated - use /voice/inbound)
      operationId: voiceCaptureGet
      deprecated: true
      parameters:
        - name: step
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: TwiML response
          content:
            application/xml:
              schema:
                type: string
        '503':
          description: Voice features not available
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT token from authenticated user session.
        Get the token from `supabase.auth.getSession()` in the frontend.
        Include in header: `Authorization: Bearer <token>`
...